requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-core
langgraph
python-dotenv
openai
requests
pydantic
python-multipart
